Separate out impl directory (makes API more clear).
