+++ packages/echo/echo-db/src/database/entity.ts
@@ -7,7 +7,7 @@ import { ItemID, ItemType } from '@dxos/echo-protocol';
 import { Model, ModelMeta, StateManager } from '@dxos/model-factory';
 import { SubscriptionGroup } from '@dxos/util';
 
-import { ItemManager } from './item-manager';
+import { ItemManager } from './impl';
 
 /**
  * Base class for all ECHO entitities.
