+++ packages/echo/echo-db/src/database/impl/data-mirror.test.ts
@@ -11,10 +11,10 @@ import { EchoEnvelope, MockFeedWriter, Timeframe } from '@dxos/echo-protocol';
 import { ModelFactory } from '@dxos/model-factory';
 import { ObjectModel } from '@dxos/object-model';
 
+import { Item } from '../item';
 import { DataMirror } from './data-mirror';
 import { DataServiceHost } from './data-service-host';
 import { DataServiceRouter } from './data-service-router';
-import { Item } from './item';
 import { ItemDemuxer } from './item-demuxer';
 import { ItemManager } from './item-manager';
