+++ packages/echo/echo-db/src/database/impl/data-service-router.ts
@@ -19,7 +19,7 @@ import {
 } from '@dxos/echo-protocol';
 import { ComplexMap } from '@dxos/util';
 
-import { PartyNotFoundError } from '../errors';
+import { PartyNotFoundError } from '../../errors';
 import { DataServiceHost } from './data-service-host';
 
 const log = debug('dxos:echo-db:data-service-router');
