+++ packages/echo/echo-db/src/database/index.ts
@@ -3,12 +3,9 @@
 //
 
 export * from './database';
-export * from './database-backend';
 export * from './entity';
+export * from './impl';
 export * from './item';
 export * from './link';
-export * from './item-demuxer';
-export * from './item-manager';
-export * from './timeframe-clock';
 export * from './selection';
 export * from './testing';
